Rename Array_reversal to Array_rev
Given an integer array arr and a list of operations operations, each operation specifies a segment of arr to reverse. The operations list is a 2D integer array where each element operations[i] contains two integers:

operations[i][0]: the starting index of the segment (inclusive).
operations[i][1]: the ending index of the segment (inclusive).
Your task is to reverse the specified segments of the array in the order they are given in operations and return the modified array.